updated agents page
